# server/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Scout Camp API
  description: Backend API for the Digital Scout Camp festival management system.
  version: 1.0.0
  contact:
    name: Digital Scout Camp Team
servers:
  - url: /api/v1
    description: Versioned API
  - url: /api
    description: Backward-compatible alias
paths:
...
  /leaderboard:
    get:
      summary: Anonymous leaderboard
      tags: [Leaderboard]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /competitions:
    get:
      summary: List competitions
      tags: [Competitions]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated competitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
...
  /reports:
    get:
      summary: List my reports
      tags: [Reports]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
    post:
      summary: Upload a report
      tags: [Reports]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
                competitionId: { type: string }
                fileName: { type: string }
                fileBase64: { type: string }
      responses:
        '201':
          description: Report uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  report: { type: object }
  /agenda:
    get:
      summary: Festival agenda
      tags: [Agenda]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Agenda items
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
...
  /news:
    get:
      summary: List news
      tags: [News]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated news
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginatedResponse:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { type: object } }
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
...